Implement 1 minute cache logic

+++ app/src/main/java/dev/maraz/weatherornot/data/DefaultWeatherRepository.kt
@@ -1,8 +1,5 @@
 package dev.maraz.weatherornot.data
 
-import androidx.lifecycle.LiveData
-import androidx.lifecycle.liveData
-import androidx.lifecycle.map
 import dev.maraz.weatherornot.data.source.local.WeatherLocalDataSource
 import dev.maraz.weatherornot.data.source.local.model.toDbModels
 import dev.maraz.weatherornot.data.source.local.model.toDomainModels
@@ -10,6 +7,10 @@ import dev.maraz.weatherornot.data.source.remote.WeatherRemoteDataSource
 import dev.maraz.weatherornot.data.source.remote.model.toDomainModel
 import dev.maraz.weatherornot.domain.model.WeatherCastData
 import kotlinx.coroutines.flow.MutableStateFlow
+import kotlinx.coroutines.flow.filterNotNull
+import kotlinx.coroutines.flow.firstOrNull
+import kotlinx.coroutines.flow.map
+import java.time.Duration
 import java.time.LocalDateTime
 import javax.inject.Inject
 import javax.inject.Singleton
@@ -20,31 +21,54 @@ class DefaultWeatherRepository @Inject constructor(
     private val weatherLocalDataSource: WeatherLocalDataSource
 ) : WeatherRepository {
 
-    private val woeid = 804365L
+    companion object {
+        private const val woeid = 804365L
+        private val cacheValidAgeThreshold = Duration.ofMinutes(1)
+    }
 
     override val isLoadingFromNetwork = MutableStateFlow(false)
 
-    override fun getCurrentWeather(): LiveData<List<WeatherCastData>?> {
-        return liveData {
-            val localData = weatherLocalDataSource.getWeatherCastData(woeid)
-                .map {
-                    it.toDomainModels()
+    // TODO publish network errors
+
+    private var latestLocalData: List<WeatherCastData>? = null
+
+    override fun getCurrentWeather() =
+        getCurrentWeatherNullable().filterNotNull()
+
+    private fun getCurrentWeatherNullable() =
+        weatherLocalDataSource.getWeatherCastData(woeid)
+            .map {
+                val domainModel = it.toDomainModels()
+                latestLocalData = domainModel
+                domainModel
+            }
+
+    override suspend fun refreshFromNetworkIfCacheHasExpired(forceRefresh: Boolean) {
+        if (forceRefresh || hasCacheExpired()) {
+            val wasLoading = !isLoadingFromNetwork.compareAndSet(expect = false, update = true)
+            if (wasLoading)
+                return
+            val result = weatherRemoteDataSource.getWeatherByLocation(woeid)
+                .map { response ->
+                    val now = LocalDateTime.now()
+                    response.toDomainModel(now)
+                }.onSuccess {
+                    weatherLocalDataSource.deleteAllWeatherCastData()
+                    weatherLocalDataSource.saveWeatherCastData(it.toDbModels())
                 }
-            emitSource(localData)
-            refreshFromRemoteAndSaveLocally(true)
+            // TODO handle network errors
+            isLoadingFromNetwork.value = false
         }
     }
 
-    override suspend fun refreshFromRemoteAndSaveLocally(forceRefresh: Boolean) {
-        isLoadingFromNetwork.value = true
-        val result = weatherRemoteDataSource.getWeatherByLocation(woeid)
-            .map { response ->
-                response.toDomainModel(LocalDateTime.now())
-            }.onSuccess {
-                weatherLocalDataSource.deleteAllWeatherCastData()
-                weatherLocalDataSource.saveWeatherCastData(it.toDbModels())
-            }
-        isLoadingFromNetwork.value = false
+    private suspend inline fun hasCacheExpired(): Boolean {
+        if (latestLocalData == null)
+            latestLocalData = getCurrentWeatherNullable().firstOrNull()
+        return latestLocalData?.firstOrNull()?.run {
+            val now = LocalDateTime.now()
+            val age = Duration.between(timestampUpdated, now)
+            age > cacheValidAgeThreshold
+        } ?: true
     }
 
 }
